test(router-ports(iridium)): iridium tested version for opmode change
Software version with opmode change working command for localizer and drifter. Using Serial Communication for Localizer and Iridium Communication with Drifter. Updated PINS/Serial handlers for both devices to make them compatible.

HasRingAsserted() callback from sbd_modem worked once at the beginning and has since then stopped working. It is not ideal to continuously check on received messages trying to read, since it consumes credits when sending a NULL message, which due to library implementation is not avoidable when reading.